runpod-serverless/requirements.txt
# Standard requirements
runpod
fastapi[all]
uvicorn[standard]
# Specific versions for compatibility
# heartlib strict dependencies - forcing available versions
numpy>=1.26.4
torch>=2.4; platform_system != 'Darwin'
torchaudio>=2.4; platform_system != 'Darwin'
transformers>=4.44.0; python_version >= '3.10'
accelerate>=1.10.1; python_version >= '3.10'
torchtune==0.4.0; platform_system != 'Darwin' and python_version >= '3.10'
torchao==0.9.0; platform_system != 'Darwin' and python_version >= '3.10'
torchvision>=0.19; platform_system != 'Darwin' and python_version >= '3.10'
# bitsandbytes 0.49.0 requires Python >=3.10 and Linux/Windows CUDA
bitsandbytes>=0.42.0; platform_system != 'Darwin' and python_version >= '3.10'
vector-quantize-pytorch==1.27.15
modelscope==1.33.0; python_version >= '3.10'
einops==0.8.1
tqdm==4.67.1
# Added more heartlib dependencies to ensure they're present when installed with --no-deps
traitlets==5.7.1
traittypes==0.2.3
tokenizers>=0.20.0
ipykernel>=6.17.1
huggingface-hub[hf_xet]
hf-transfer
sentencepiece
protobuf
scipy
basic-pitch; python_version >= '3.10'
pretty_midi
librosa>=0.11.0
soundfile
mido
pyyaml
python-multipart
python-dotenv
passlib[bcrypt]
pydantic
websockets
anyio
